Created a separate dip_filter module. Moved calling of functions into the __main__.py script.
